fix: MSVC 2026 の C4242 警告を解消（縮小変換の明示化）
VS2026 (PlatformToolset v145) で /Wall + /WX により C4242
(possible loss of data) がエラー化したため、暗黙の縮小変換に
static_cast を追加。

主な変換パターン:
- int → ARMOUR_CLASS (int16_t): get_ac() 戻り値を test_hit_norm() や
  ローカル ac 等に渡す箇所
- int → MONSTER_IDX (int16_t): floor.get_monster(int) 呼出
- int → PLAYER_LEVEL (int16_t): get_level() 戻り値の代入
- int → byte: speed_to_energy() の引数（get_speed() 戻り値）
- int → byte: CapturedMonsterType.speed = int speed
- int → SUB_EXP (int16_t): skill_exp 加算
- int → PARAMETER_VALUE (int16_t): item.pval 代入
- int → int16_t: DeathRecord.day/hour/min, old_speed 等
- int → uint32_t: s64b_mul の 4 引数目
- int → byte: wr_byte(enum2i(...))
- int → int16_t: wr_s16b(enum2i(...) / item.captured_monster_max_hp)

42 ファイル 57 箇所を修正。

+++ src/birth/birth-select-patron.cpp
@@ -167,7 +167,7 @@ bool get_player_patron(CreatureEntity &creature)
         return false;
     }
 
-    creature.patron = k;
+    creature.patron = static_cast<int16_t>(k);
     display_player_name(creature);
     return true;
 }
